create CityAdmin

+++ farmer_market/models.py
@@ -4,8 +4,8 @@ from django.db import models
 class City(models.Model):
     name = models.CharField(max_length=25)
     dept = models.SmallIntegerField()
-    long = models.SmallIntegerField()
-    lat = models.SmallIntegerField()
+    long = models.FloatField()
+    lat = models.FloatField()
 
     class Meta:
         verbose_name = "city"
